refactoring without Service Manager
