+++ README.md
@@ -16,15 +16,12 @@ composer require baywa-re-lusy/publisher-subscriber
 Currently, this library only supports PubNub. However, it uses an Adapter pattern to allow adding other vendors easily.
 
 ```php
-use BayWaReLusy\PublisherSubscriberTools\PubSubTools;
-use BayWaReLusy\PublisherSubscriberTools\PubSubToolsConfig;
-use BayWaReLusy\PublisherSubscriberTools\PubSubService;
-use BayWaReLusy\PublisherSubscriberTools\Adapter\PubNubAdapter;
+use BayWaReLusy\PublisherSubscriber\PubNubService;use BayWaReLusy\PublisherSubscriber\PubSubService;use BayWaReLusy\PublisherSubscriber\PubSubTools;use BayWaReLusy\PublisherSubscriber\PubSubToolsConfig;
 
 $pubSubToolsConfig = new PubSubToolsConfig('publisher-key', 'subscriber-key');
 $pubSubTools       = new PubSubTools($pubSubToolsConfig);
 $pubSubService     = $pubSubTools->get(PubSubService::class);
-$pubSubService->setAdapter($pubSubTools->get(PubNubAdapter::class));
+$pubSubService->setAdapter($pubSubTools->get(PubNubService::class));
 ```
 
 Optionally, you can include then the PubSub Client into your Service Manager:
