+++ composer.json
@@ -5,17 +5,16 @@
     "license": "proprietary",
     "homepage": "https://github.com/baywa-re-lusy/publisher-subscriber",
     "require": {
-        "php": ">= 7.4 || >= 8.0",
-        "laminas/laminas-servicemanager": "^3.6",
-        "pubnub/pubnub": "^4.6"
+        "php": ">= 8.1",
+        "pubnub/pubnub": "^5.1"
     },
     "require-dev": {
         "squizlabs/php_codesniffer": "^3.5",
-        "phpstan/phpstan": "^1.6"
+        "phpstan/phpstan": "^1.9"
     },
     "autoload": {
         "psr-4": {
-            "BayWaReLusy\\PublisherSubscriberTools\\": "src/PublisherSubscriberTools"
+            "BayWaReLusy\\PublisherSubscriber\\": "src/PublisherSubscriber"
         }
     }
 }
